# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Returns `data.access_token`. Send that token back in the `authorization` header exactly as received -- the guard verifies the raw header value, so do NOT add a \"Bearer \" prefix. Wrong credentials answer 200 with `errorCode` `1000`.",
        "operationId": "ExternalAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "Log in and obtain a JWT",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/detail": {
      "post": {
        "description": "Decodes the JWT in the `authorization` header and returns its claims.",
        "operationId": "ExternalAuthController_userDetail",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Current user detail",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a `Customer` user. When `referal_code` is supplied the customer is also attached to the owning admin. Duplicate usernames answer with `errorCode` `1003`.",
        "operationId": "ExternalAuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a customer account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/admin/api-key": {
      "post": {
        "description": "Returns a freshly generated API key in `data.apiKey`. It is only shown once -- the stored copy is encrypted. Use it to sign request bodies for the `x-sign` header.",
        "operationId": "ExternalAuthController_reqApiKeyAdmin",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Generate an API key for the calling admin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/api-key": {
      "post": {
        "description": "Same as `/admin/api-key`, but the target user is named in the body.",
        "operationId": "ExternalAuthController_reqApiKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Generate an API key for another user (SuperAdmin)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/generate-referal": {
      "post": {
        "description": "Admin only. Creates the referral code customers use at registration.",
        "operationId": "ExternalAuthController_reqReferal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReferalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Generate a referral code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/check-referal": {
      "post": {
        "description": "Public. Resolves a referral code to its owning admin. `1005` when unknown.",
        "operationId": "ExternalAuthController_checkReferal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckReferalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check a referral code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/assign-admin": {
      "post": {
        "description": "Customer only. `1006` when the customer is already assigned.",
        "operationId": "ExternalAuthController_assignToAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Attach the current customer to an admin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/customer-list": {
      "post": {
        "description": "Paginated. `name` narrows the result set.",
        "operationId": "ExternalAuthController_getAdminCustomers",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCustomerListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the customers of the calling admin",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "TRX API Service",
    "description": "Transaction / top-up API service.\n\n## Authentication\n\nCall `POST /api/v1/auth/login` to obtain a JWT, then send it in the\n`authorization` header **without** the `Bearer ` prefix -- the guard verifies\nthe header value verbatim.\n\nMachine-to-machine endpoints (those under `/api/v1/tx/ext/...`) accept the\n`x-username` + `x-sign` header pair instead of a JWT. `x-sign` is the\nHMAC-SHA256 hex digest of the exact JSON request body signed with the\ncaller's API key (see `POST /api/v1/auth/api-key`).\n\nA few endpoints (`/api/v1/tx/request-web`, `/api/v1/external-topup/create-topup`,\n`/api/v1/external-topup/update-topup`) require BOTH a JWT and an `x-sign`\nheader over the body.\n\n## Response envelope\n\nMost endpoints answer with `{ message, data, errorCode }`. An `errorCode` of\n`\"0000\"` means success; any other code maps to an entry in\n`src/lib/errorCodes.ts` (e.g. `1000` wrong credential, `1002` insufficient\nbalance, `4012` signature not verified).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Raw JWT returned by POST /api/v1/auth/login. Do NOT prefix it with \"Bearer \"."
      },
      "x-username": {
        "type": "apiKey",
        "in": "header",
        "name": "x-username",
        "description": "Username of the API consumer. Used together with x-sign."
      },
      "x-sign": {
        "type": "apiKey",
        "in": "header",
        "name": "x-sign",
        "description": "HMAC-SHA256 hex signature of the JSON request body."
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "success",
            "description": "Human readable result message"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Endpoint specific payload. `null` when the call failed."
          },
          "errorCode": {
            "type": "string",
            "example": "0000",
            "description": "`0000` on success, otherwise a code from src/lib/errorCodes.ts"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "example": "S3cretPassw0rd",
            "format": "password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginResultDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT to send back in the `authorization` header, verbatim (no \"Bearer \" prefix).",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJqb2huZG9lIn0.sig"
          }
        },
        "required": [
          "access_token"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LoginResultDto"
              }
            ]
          },
          "errorCode": {
            "type": "string",
            "example": "0000",
            "description": "`1000` when the credentials are wrong"
          }
        },
        "required": [
          "message",
          "data",
          "errorCode"
        ]
      },
      "SignUpDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "example": "S3cretPassw0rd",
            "format": "password"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "phone_number": {
            "type": "string",
            "example": "628123456789"
          },
          "referal_code": {
            "type": "string",
            "description": "Referral code of the admin this customer should be attached to.",
            "example": "2f6b3f7a-4d1c-4b2e-9f0a-7c1d2e3f4a5b"
          }
        },
        "required": [
          "username",
          "name",
          "password"
        ]
      },
      "GenerateApiKeyDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "SuperAdmin only: username to generate the key for. Omit to generate it for the caller.",
            "example": "johndoe"
          }
        }
      },
      "GenerateReferalDto": {
        "type": "object",
        "properties": {
          "referal_name": {
            "type": "string",
            "example": "reseller-jakarta",
            "description": "Display name of the referral"
          }
        },
        "required": [
          "referal_name"
        ]
      },
      "CheckReferalDto": {
        "type": "object",
        "properties": {
          "referal_code": {
            "type": "string",
            "example": "2f6b3f7a-4d1c-4b2e-9f0a-7c1d2e3f4a5b"
          }
        },
        "required": [
          "referal_code"
        ]
      },
      "AssignAdminDto": {
        "type": "object",
        "properties": {
          "referal_code": {
            "type": "string",
            "description": "Referral code of the admin the current customer should be assigned to.",
            "example": "2f6b3f7a-4d1c-4b2e-9f0a-7c1d2e3f4a5b"
          }
        },
        "required": [
          "referal_code"
        ]
      },
      "AdminCustomerListDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "size": {
            "type": "number",
            "example": 10
          },
          "name": {
            "type": "string",
            "description": "Filter by customer name",
            "example": "john"
          }
        },
        "required": [
          "page",
          "size"
        ]
      },
...
    }
  }
}